requirements.txt
mysql-connector-python==8.2.0
python-logging-loki==0.3.1
charset-normalizer==3.3.2
python-dateutil==2.8.2
python-dotenv==1.0.0
certifi==2023.11.17
requests==2.31.0
schedule==1.2.1
pysmb==1.2.9.1
urllib3==2.1.0
pyotp==2.9.0
rfc3339==6.2
six==1.16.0
idna==3.6
